requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
greenlet==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
